move GameProgress
