+++ frontend/src/game/game.js
@@ -3,72 +3,8 @@ import {Component} from "react";
 import {Users} from "../users/users";
 import {Username} from "../authentication/authentication";
 import {GameSetup} from "./game_start.js"
+import {GameProgress} from "./game_progress.js"
 import {StartRound} from "./round_start.js"
-
-let roundHistory = [
-    1,
-    2,
-    1,
-    0,
-    null,
-    null,
-    null,
-];
-
-export function GameProgress() {
-    return (
-        <div className="GameProgress">
-            <table>
-                <tr>
-                    <td>
-                        Round
-                    </td>
-                    {roundHistory.map((roundStatus, roundIndex) => {
-                        return (
-                            <td>
-                                {roundIndex + 1}
-                            </td>
-                        )
-                    })}
-                </tr>
-                <tr>
-                    <td>
-                        Status
-                    </td>
-                    {roundHistory.map((roundStatus, roundIndex) => {
-                        switch (roundStatus) {
-                            case null:
-                                return (
-                                    <td>
-
-                                    </td>
-                                );
-                            case 0:
-                                return (
-                                    <td>
-                                        ?
-                                    </td>
-                                );
-                            case 1:
-                                return (
-                                    <td style={{color: "green"}}>
-                                        W
-                                    </td>
-                                );
-                            case 2:
-                                return (
-                                    <td style={{color: "red"}}>
-                                        L
-                                    </td>
-                                );
-
-                        }
-                    })}
-                </tr>
-            </table>
-        </div>
-    )
-}
 
 
 export class GamePreparation extends Component {
